% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getCohortMethodResult}
\alias{getCohortMethodResult}
\title{Get the main study results data.}
\usage{
getCohortMethodResult(
  connection = NULL,
  databaseSchema = NULL,
  cohortMethodResult = NULL,
  cohortMethodAnalysis = NULL,
  targetId,
  comparatorId,
  outcomeId,
  databaseId,
  analysisId
)
}
\arguments{
\item{connection}{(optional) An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{databaseSchema}{(optional) The databaseSchema argument is interpreted differently according
to the different supported RDMS platforms: SQLServer and PDW: The databaseSchema
schema should specify both the database and the schema,e.g. ’my_database.dbo’.
PostgreSQL and Redshift: The databaseSchema should specify the schema.
Oracle:  The databaseSchema should specify the Oracle ’user’.}

\item{cohortMethodResult}{(optional) A R-dataFrame object with fields named in camelCase containing
data from cohort_method_result table in comparative effectiveness data model.}

\item{cohortMethodAnalysis}{(optional) A R-dataFrame object with fields named in camelCase containing
data from cohort_method_analysis table in comparative effectiveness data model.}
}
\value{
Tibble. Fields with column names 'HR (95% CI)', 'Cal. HR (95% CI)' are provided
                            in print friendly format.
}
\description{
Get the main study results data.
}
\details{
Get the main results data for target, comparator and outcome cohorts in data frame format from
OHDSI comparative effectiveness data model. The output maybe used to create plots or tables
that use the main study results data.
}
\examples{
\dontrun{
#If your data is in database, provide connection
connection <- connect(createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
                                             )
example1 <- getCohortMethodResult(connection = connection,
                                  databaseSchema = 'database.schema',
                                  targetId = 100,
                                  comparatorId = 200,
                                  outcomeId = 111,
                                  databaseId = 'HCUP data',
                                  analysisId = 2)
example2 <- getCohortMethodResult(cohortMethodResult = cohortMethodResult,
                                  cohortMethodAnalysis = cohortMethodAnalysis,
                                  targetId = 100,
                                  comparatorId = 200,
                                  outcomeId = 111,
                                  databaseId = 'HCUP data',
                                  analysisId = 2)
}
}
