% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getCovariateBalance}
\alias{getCovariateBalance}
\title{Get covariate balance data.}
\usage{
getCovariateBalance(
  connection = NULL,
  databaseSchema = NULL,
  covariateBalance = NULL,
  covariate = NULL,
  targetId,
  comparatorId,
  outcomeId,
  databaseId,
  analysisId,
  interactionCovariateId = NULL
)
}
\arguments{
\item{connection}{(optional) An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{databaseSchema}{(optional) The databaseSchema argument is interpreted differently according
to the different supported RDMS platforms: SQLServer and PDW: The databaseSchema
schema should specify both the database and the schema,e.g. ’my_database.dbo’.
PostgreSQL and Redshift: The databaseSchema should specify the schema.
Oracle:  The databaseSchema should specify the Oracle ’user’.}

\item{covariateBalance}{(optional) A R-dataFrame object with fields named in camelCase containing
data from covariate_balance table in comparative effectiveness data model.}

\item{covariate}{(optional) A R-dataFrame object with fields named in camelCase containing
data from covariate table in comparative effectiveness data model.}

\item{targetId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{comparatorId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{outcomeId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{databaseId}{A short string for identifying the database for query from comparative effectiveness
data model (e.g. 'Synpuf')}

\item{analysisId}{An integer identifying the analysis_id field in cohort_method_analysis table
of OHDSI comparative effectiveness data model}

\item{interactionCovariateId}{(optoinal) An integer identifying the covariate in interactionCovariateId
field of covariate_balance table of OHDSI comparative effectiveness data model}
}
\value{
Tibble.
}
\description{
Get covariate balance data.
}
\details{
Get covariate balance data from data stored in OHDSI comparative effectiveness data model,
The output of this function  maybe used to create plots or tables
that uses the covariate balance data.

If the comparative effectiveness results are stored in a
relational database server, then please provide connection details
using DatabaseConnector::connection. If connection
parameter is not null, the function will attempt to connect to
relational database server, and ignore using R dataframe object.

If the comparative effectiveness results are stored as a R dataframe object,
please make sure connection is NULL (Default value). The function
will expect data in R dataframe. This R dataframe object is expected to follow
camelCase for both object names and variable names.
}
\examples{
\dontrun{
}
}
