% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getFollowUpDistribution}
\alias{getFollowUpDistribution}
\title{Get followup distribution data.}
\usage{
getFollowUpDistribution(
  connection = NULL,
  databaseSchema = NULL,
  cmFollowUpDist = NULL,
  targetId,
  comparatorId,
  outcomeId,
  databaseId,
  analysisId
)
}
\arguments{
\item{connection}{(optional) An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{cmFollowUpDist}{(optional) A R-dataFrame object with fields named in camelCase containing
data from cm_follow_up_dist table in comparative effectiveness data model.}

\item{targetId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{comparatorId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{outcomeId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{databaseId}{A short string for identifying the database for query from comparative effectiveness
data model (e.g. 'Synpuf')}

\item{analysisId}{An integer identifying the analysis_id field in cohort_method_analysis table
of OHDSI comparative effectiveness data model}
}
\value{
Tibble with fields Id, Min, P10, P25, Median, P75, P90,Max, group.
                            group is for identifying if the cohort is target vs comparator.
}
\description{
Get followup distribution data.
}
\details{
Get follow-up distribution data from data stored in OHDSI comparative effectiveness data model,
The output of this function  maybe used to create plots or tables
that uses the follow-up distribution data.

If the comparative effectiveness results are stored in a
relational database server, then please provide connection details
using DatabaseConnector::connection. If connection
parameter is not null, the function will attempt to connect to
relational database server, and ignore using R dataframe object.

If the comparative effectiveness results are stored as a R dataframe object,
please make sure connection is NULL (Default value). The function
will expect data in R dataframe. This R dataframe object is expected to follow
camelCase for both object names and variable names.
}
\examples{
\dontrun{
#If your data is in database, provide connection
connection <- connect(createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
                                             )
#If connection is provided, function will attempt to read data from database
example1 <- getFollowUpDaysDistribution(connection = connection,
                                        databaseSchema = 'database.schema',
                                        targetId = 100,
                                        comparatorId = 200,
                                        outcomeId = 111,
                                        databaseId = 'HCUP data',
                                        analysisId = 2)
#If no connection is provided, then function will check for data in dataFrame
example2 <- getFollowUpDaysDistribution(cmFollowUpDist = cmFollowUpDist,
                                        targetId = 100,
                                        comparatorId = 200,
                                        outcomeId = 111,
                                        databaseId = 'HCUP data',
                                        analysisId = 2)
}
}
