% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getMetadataForNaturalKeys}
\alias{getMetadataForNaturalKeys}
\title{Get metadata on models natural keys}
\usage{
getMetadataForNaturalKeys(
  connection = NULL,
  databaseSchema = NULL,
  cohortMethodAnalysis = NULL,
  covariateAnalysis = NULL,
  exposureOfInterest = NULL,
  outcomeOfInterest = NULL,
  negativeControlOutcome = NULL,
  positiveControlOutcome = NULL,
  database = NULL
)
}
\arguments{
\item{connection}{(optional) An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{databaseSchema}{(optional) The databaseSchema argument is interpreted differently according
to the different supported RDMS platforms: SQLServer and PDW: The databaseSchema
schema should specify both the database and the schema,e.g. ’my_database.dbo’.
PostgreSQL and Redshift: The databaseSchema should specify the schema.
Oracle:  The databaseSchema should specify the Oracle ’user’.}

\item{cohortMethodAnalysis}{(optional) A R-dataFrame object with fields named in camelCase containing
data from cohort_method_analysis table in comparative effectiveness data model.}

\item{covariateAnalysis}{(optional) A R-dataFrame object with fields named in camelCase containing
data from covariate_analysis table in comparative effectiveness data model.}

\item{exposureOfInterest}{(optional) A R-dataFrame object with fields named in camelCase containing
data from exposure_of_interest table in comparative effectiveness data model.}

\item{outcomeOfInterest}{(optional) A R-dataFrame object with fields named in camelCase containing
data from outcome_of_interest table in comparative effectiveness data model.}

\item{negativeControlOutcome}{(optional) A R-dataFrame object with fields named in camelCase containing
data from negative_control_outcome table in comparative effectiveness data model.}

\item{positiveControlOutcome}{(optional) A R-dataFrame object with fields named in camelCase containing
data from positive_control_outcome table in comparative effectiveness data model.}

\item{database}{(optional) A R-dataFrame object with fields named in camelCase containing
data from database table in comparative effectiveness data model.}
}
\value{
Tibble data frame with fields table, key, keyId, keyName, keyDescription.
}
\description{
Get metadata on models natural keys
}
\details{
Get the metadata on natural keys used in the comparative effectiveness data model tables
cohort_method_analysis, covariate_analysis, exposure_of_interest, outcome_of_interest,
negative_control_outcome, positive_control_outcome, database.
}
\examples{
\dontrun{

#If your data is in database, provide connection.
connection <- connect(createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
                                             )
#If connection is provided, function will attempt to read data from database
# data in R-dataframe format will be ignored
example1 <- getMetadataForNaturalKeys(connection = NULL)
#If no connection is provided, then function will check for data in dataFrame
example2 <- getMetadataForNaturalKeys(connection = NULL,
                                      cohortMethodAnalysis = cohortMethodAnalysis,
                                      covariateAnalysis = covariateAnalysis,
                                      exposureOfInterest = exposureOfInterest)
}
}
