% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getPower}
\alias{getPower}
\title{Get Power data}
\usage{
getPower(
  connection = NULL,
  databaseSchema = NULL,
  cohortMethodResult = NULL,
  cohortMethodAnalysis = NULL,
  targetId,
  comparatorId,
  outcomeId,
  databaseId,
  analysisId
)
}
\arguments{
\item{connection}{(optional) An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{databaseSchema}{(optional) The databaseSchema argument is interpreted differently according
to the different supported RDMS platforms: SQLServer and PDW: The databaseSchema
schema should specify both the database and the schema,e.g. ’my_database.dbo’.
PostgreSQL and Redshift: The databaseSchema should specify the schema.
Oracle:  The databaseSchema should specify the Oracle ’user’.}

\item{cohortMethodResult}{(optional) A R-dataFrame object with fields named in camelCase containing
data from cohort_method_result table in comparative effectiveness data model.}

\item{cohortMethodAnalysis}{(optional) A R-dataFrame object with fields named in camelCase containing
data from cohort_method_analysis table in comparative effectiveness data model.}

\item{targetId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{comparatorId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{outcomeId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{databaseId}{A short string for identifying the database for query from comparative effectiveness
data model (e.g. 'Synpuf')}

\item{analysisId}{An integer identifying the analysis_id field in cohort_method_analysis table
of OHDSI comparative effectiveness data model}
}
\value{
Tibble.
}
\description{
Get Power data
}
\details{
Get the power results for target, comparator and outcome cohorts in data frame format from
OHDSI comparative effectiveness data model. The output maybe used to create plots or tables.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                           user = "joe",
                                           password = "secret",
                                           server = "myserver")
example1 <- getPower(connection = connectionDetails,
                     databaseSchema = 'database.schema',
                     targetId = 100,
                     comparatorId = 200,
                     outcomeId = 111,
                     databaseId = 'HCUP data',
                     analysisId = 2)
example2 <- getPower(cohortMethodResult = cohortMethodResult,
                     cohortMethodAnalysis = cohortMethodAnalysis,
                     targetId = 100,
                     comparatorId = 200,
                     outcomeId = 111,
                     databaseId = 'HCUP data',
                     analysisId = 2)
}
}
