% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getPreferenceScore}
\alias{getPreferenceScore}
\title{gets a table with data from preference score distribution.}
\usage{
getPreferenceScore(
  connectionDetails = NULL,
  preferenceScoreDist = NULL,
  exposureOfInterest = NULL,
  targetId,
  comparatorId,
  databaseId,
  analysisId
)
}
\arguments{
\item{connectionDetails}{(optional) An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{preferenceScoreDist}{(optional) A R-dataFrame object with fields named in camelCase containing
data from preference_score_dist table in comparative effectiveness data model.}

\item{exposureOfInterest}{(optional) A R-dataFrame object with fields named in camelCase containing
data from exposure_of_interest table in comparative effectiveness data model.}

\item{targetId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{comparatorId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{databaseId}{A short string for identifying the database for query from comparative effectiveness
data model (e.g. 'Synpuf')}

\item{analysisId}{An integer identifying the analysis_id field in cohort_method_analysis table
of OHDSI comparative effectiveness data model}
}
\value{
tibble
}
\description{
gets a table with data from preference score distribution.
}
\details{
gets preference score data for target and comparator in data frame format from
OHDSI comparative effectiveness data model. The output maybe used to create plots or tables
that compares the preference score distribution between target and comparators.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
example1 <- getPreferenceScore(connectionDetails = connectionDetails,
                               targetId = 100,
                               comparatorId = 200,
                               databaseId = 'HCUP data',
                               analysisId = 2)
example2 <- getPreferenceScore(preferenceScoreDist = preferenceScoreDist,
                               exposureOfInterest = exposureOfInterest,
                               targetId = 100,
                               comparatorId = 200,
                               databaseId = 'HCUP data',
                               analysisId = 2)
}
}
