% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getPreferenceScoreDistribution}
\alias{getPreferenceScoreDistribution}
\title{Get the preference score distribution for plotting}
\usage{
getPreferenceScoreDistribution(
  connectionDetails = NULL,
  databaseSchema = NULL,
  preferenceScoreDist = NULL,
  exposureOfInterest = NULL,
  targetId,
  comparatorId,
  databaseId,
  analysisId
)
}
\arguments{
\item{connectionDetails}{(optional) An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{databaseSchema}{(optional) The databaseSchema argument is interpreted differently according
to the different supported RDMS platforms: SQLServer and PDW: The databaseSchema
schema should specify both the database and the schema,e.g. ’my_database.dbo’.
PostgreSQL and Redshift: The databaseSchema should specify the schema.
Oracle:  The databaseSchema should specify the Oracle ’user’.}

\item{preferenceScoreDist}{(optional) A R-dataFrame object with fields named in camelCase containing
data from preference_score_dist table in comparative effectiveness data model.}

\item{exposureOfInterest}{(optional) A R-dataFrame object with fields named in camelCase containing
data from exposure_of_interest table in comparative effectiveness data model.}

\item{targetId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{comparatorId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{databaseId}{A short string for identifying the database for query from comparative effectiveness
data model (e.g. 'Synpuf')}

\item{analysisId}{An integer identifying the analysis_id field in cohort_method_analysis table
of OHDSI comparative effectiveness data model}
}
\value{
Tibble with fields preference score, density, cohort-name,
                            group for identifying if the cohort is target vs comparator.
}
\description{
Get the preference score distribution for plotting
}
\details{
Get preference score data from data stored in OHDSI comparative effectiveness data model,
The output of this function  maybe used to create plots or tables
that compares the preference score distribution between target and comparators.
}
\examples{
# example1: If the comparative effectiveness results are stored in a 
# relational database server: then please provide connection details 
# using DatabaseConnector::createConnectionDetails. 
# If connectionDetails parameter is not null, the function will attempt to
# connect to relational database server, and ignore using at R dataframe objects
# example2: If the comparative effectiveness results are stored as a R dataframe object.
# Note: R dataframes are expected to follow camelCase for both object names 
# and variable names.
# Note: schema parameter in createConnectionDetails should not be used. Use
# databaseSchema in function call instead.                  
\dontrun{


connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
example1 <- getPreferenceScore(connectionDetails = connectionDetails,
                               databaseSchema = 'hcup.version2020',
                               targetId = 100,
                               comparatorId = 200,
                               databaseId = 'HCUP data',
                               analysisId = 2)
example2 <- getPreferenceScore(preferenceScoreDist = preferenceScoreDist,
                               exposureOfInterest = exposureOfInterest,
                               targetId = 100,
                               comparatorId = 200,
                               databaseId = 'HCUP data',
                               analysisId = 2)
}
}
