% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotCovariateBalanceScatter}
\alias{plotCovariateBalanceScatter}
\title{Get the covariate balance scatter plot for data covariateBalance object.}
\usage{
plotCovariateBalanceScatter(covariateBalance = NULL, fileName = NULL)
}
\arguments{
\item{covariateBalance}{A R-dataFrame covariateBalance object that was derived by 
function \code{getCovariateBalance}}

\item{fileName}{(optional) Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{ggsave} in the ggplot2 package for supported file
formats.}
}
\value{
ggplot object. if fileName is provided, a ggsave will be used to save the object.
}
\description{
Get the covariate balance scatter plot for data covariateBalance object.
}
\details{
This function creates the covariate balance scatter plot for data in 
covariateBalance object. The covariateBalance object is created from OHDSI
comparative effectiveness data model by function \code{getCovariateBalance}. 

(Advanced) The output is a ggplot object. It maybe modified to change color, title, label etc.
}
\examples{
\dontrun{
example1 <- plotCovariateBalanceScatter(covariateBalance = covariateBalance)
} 
}
