% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{SkeletonComparativeEffectStudy-package}
\alias{SkeletonComparativeEffectStudy}
\alias{SkeletonComparativeEffectStudy-package}
\title{SkeletonComparativeEffectStudy: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
