% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{synthesizeResults}
\alias{synthesizeResults}
\title{Title}
\usage{
synthesizeResults(
  allDbsFolder,
  maExportFolder = allDbsFolder,
  maxCores,
  method = "bayesian",
  resultsZipPattern = "^Results_.*\\\\.zip",
  addTraditional = TRUE
)
}
\arguments{
\item{allDbsFolder}{Folder on the local file system containing the individual zip files across databases (i.e., sites)}

\item{maExportFolder}{A local folder where the meta-analysis results will be written. If not specified, results will be
written to same directory with all other results.}

\item{maxCores}{Maximum number of CPU cores to be used when computing the meta-analyses.}

\item{method}{The meta-analysis method to use.  Options are "bayesian" or "DL" (DerSimonian-Laird).}

\item{addTraditional}{Boolean indicating if traditional meta-analysis (i.e., "DL") results should be added to result (if method is "bayesian").}
}
\value{
Does not return a value, but creates a new zip file in the \code{maExportFolder} for the meta-analyses.
}
\description{
Title
}
