% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPulls.R
\name{getPreferenceScoreDistribution}
\alias{getPreferenceScoreDistribution}
\title{Get the preference score distribution for plotting}
\usage{
getPreferenceScoreDistribution(
  connection = NULL,
  databaseSchema = NULL,
  preferenceScoreDist = NULL,
  exposureOfInterest = NULL,
  targetId,
  comparatorId,
  databaseId,
  analysisId
)
}
\arguments{
\item{preferenceScoreDist}{(optional) A R-dataFrame object with fields named in camelCase containing
data from preference_score_dist table in comparative effectiveness data model.}

\item{exposureOfInterest}{(optional) A R-dataFrame object with fields named in camelCase containing
data from exposure_of_interest table in comparative effectiveness data model.}

\item{targetId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{comparatorId}{An integer identifying a cohort in exposure_id field of exposure_of_interest
of OHDSI comparative effectiveness data model}

\item{databaseId}{A short string for identifying the database for query from comparative effectiveness
data model (e.g. 'Synpuf')}

\item{analysisId}{An integer identifying the analysis_id field in cohort_method_analysis table
of OHDSI comparative effectiveness data model}

\item{optionalConnection}{(optional) An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}
}
\value{
Tibble with fields preference score, density, cohort-name, group.
                            group is for identifying if the cohort is target vs comparator.
}
\description{
Get the preference score distribution for plotting
}
\details{
Get preference score data from data stored in OHDSI comparative effectiveness data model,
The output of this function  maybe used to create plots or tables
that compares the preference score distribution between target and comparators.

If the comparative effectiveness results are stored in a
relational database server, then please provide connection
using DatabaseConnector::connect. If connection
parameter is not null, the function will attempt to connect to
relational database server, and ignore using R dataframe object.


If the comparative effectiveness results are stored as a R dataframe object,
please make sure connection is NULL (Default value). The function
will expect data in R dataframe. This R dataframe object is expected to follow
camelCase for both object names and variable names.
}
\examples{
\dontrun{

#If your data is in database, provide connection.
connection <- connect(createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
                                             )
#If connection is provided, function will attempt to read data from database
# data in R-dataframe format will be ignored
example1 <- getPreferenceScore(connection = connection,
                               databaseSchema = 'hcup.version2020',
                               targetId = 100,
                               comparatorId = 200,
                               databaseId = 'HCUP data',
                               analysisId = 2)
#If no connection is provided, then function will check for data in dataFrame
example2 <- getPreferenceScore(preferenceScoreDist = preferenceScoreDist,
                               exposureOfInterest = exposureOfInterest,
                               targetId = 100,
                               comparatorId = 200,
                               databaseId = 'HCUP data',
                               analysisId = 2)
}
}
