% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotPreferenceScore}
\alias{plotPreferenceScore}
\title{Get the preference score distribution plot for data preferenceScore object.}
\usage{
plotPreferenceScore(preferenceScore = NULL, fileName = NULL)
}
\arguments{
\item{preferenceScore}{A R-dataFrame preferenceScore object that was derived by 
function \code{getPreferenceScoreDistribution}}

\item{fileName}{(optional) Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{ggsave} in the ggplot2 package for supported file
formats.}
}
\value{
ggplot object. if fileName is provided, a ggsave will be used to save the object.
}
\description{
Get the preference score distribution plot for data preferenceScore object.
}
\details{
This function creates the preference score distribution plot for data in 
preferenceScore object. The preferenceScore object is created from OHDSI
comparative effectiveness data model by function \code{getPreferenceScoreDistribution}. 
The output of this function visualizes the preference score distribution 
between target and comparators.

(Advanced) The output is a ggplot object. It maybe modified to change color, title, label etc.
}
\examples{
\dontrun{
example1 <- plotPreferenceScore(preferenceScore = preferenceScore)
} 
}
